PaulstretchBase in lib-builtin-effects
